% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LoF.R
\name{compute_LoF}
\alias{compute_LoF}
\title{Compute Loss of Function gene}
\usage{
compute_LoF(df)
}
\arguments{
\item{df}{Data frame containing variants}
}
\value{
Ramka danych z dodanymi kolumnami LoF
}
\description{
Compute Loss of Function gene
}
