% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHDetect.R
\name{MHDetect}
\alias{MHDetect}
\title{MHDetect Function}
\usage{
MHDetect(vcf, k, N, genome, Interval)
}
\arguments{
\item{vcf}{A VCF file (Variant Call Format) object read with readVcf.}

\item{k}{Number of base pairs to analyze before and after a deletion.}

\item{N}{Minimum number of matching nucleotides to classify as MMEJ.}

\item{genome}{The genome reference, e.g., BSgenome.Hsapiens.UCSC.hg19.}

\item{Interval}{Length of the region for analyzing DNA repair.}
}
\value{
A data frame with classified indels and additional features.
}
\description{
This function detects and classifies indels (deletions, insertions, and MNVs)
as microhomology-mediated end joining (MMEJ)-dependent.
}
\examples{
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)
strelka <- "path/to/your/file.vcf.gz"
vcf_data <- readVcf(strelka, genome = "hg19")
result <- MHDetect(vcf_data, k = 25, N = 2,
                   genome = BSgenome.Hsapiens.UCSC.hg19, Interval = 25)
}
}
